Shadin backend

+++ lms-backend/src/main/java/com/lms/loan/dto/BusinessLoanDetailsDto.java
@@ -2,13 +2,15 @@ package com.lms.loan.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class BusinessLoanDetailsDto {
 
     private String businessName;
     private String businessType; // MSME / Retail / etc.
 
-    private Double gstAnnualTurnover;
+    private BigDecimal gstAnnualTurnover;
     private Integer businessVintageYears;
 
     private String proofOfBusiness;
