+++ lms-backend/src/main/java/com/lms/loan/dto/PersonalLoanDetailsDto.java
@@ -2,11 +2,13 @@ package com.lms.loan.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class PersonalLoanDetailsDto {
 
     private String employmentType; // SALARIED / SELF_EMPLOYED
-    private Double monthlyIncome;
+    private BigDecimal monthlyIncome;
     private String employerName;
 
     private String proofOfIdentity;
